Pybindings for I3HDFTableService + a sketch of a user interface.
1. Added pybindings for I3HDFTableService. I think we should make a clean
   break and force the user to pass an already-constructed service, v3 style. 
   No more factories!

2. Began to sketch out the user interface. I think it's best to write the
   I3Module itself in Python. That lets us do a few slick things, such as the
   following interface.

The beginnings of a prototype I3Module are in resources/scripts/module.py

There are 3 parameters: TableService, Keys, and Types. TableService is
required, and must be an instance of I3TableService.

In its most advanced form, Types is a list of (boost-python-type,
python-wrapped-converter) tuples. The converters will have to be initialized
by the user, like services. This takes care of any parameters the converter
might have. The I3TableWriter should raise an error if the passed converter
can't handle the specified type. The tuple structure means that you can
specify multiple converters for a single type. If the types are unique, then
the user can also pass a dict, which will be converted to a list of tuples. If
the user just passes a list of types, or equivalently a {boost-python-type:
None} dict, the I3TableWriter should use the default booker.

Keys will operate similarly: multiple tuples can be passed with the same key
to run more than one converter on the frame object. Dicts and lists of keys
will be expanded accordingly. If an element of the key-list conflicts with an
element of the type-list, then the key takes precedence. For example, if the
user passes types={dataclasses.I3Particle: MyParticleBooker()} _and_
keys=['credofit'], then 'credofit' will be converted using the compiled
I3Particle converter, not MyParticleBooker.

The usual caveats apply to alignment: alignment is only strictly guaranteed if
the user specifies both the key name and the converter to use.

What is missing, I realize, is a way to specify the name of the table.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61470 16731396-06f5-0310-8873-f7f720988828