Actual, honest-to-god type comparisons, and unit/doc metadata in I3HDFTable.
Lots of fun stuff here. First off, FrameObject type inference is handled in a 
more robust way. I3TableWriter spools through the converter registry at startup 
and instantiates each registered converter. When an object is added to the 
table list, the writer hands each booker an I3FrameObjectConstPtr and asks if 
it can book it (the default implementation compares the type of the template 
argument to the passed FrameObject). The converter that answers the call gets 
the job. No more string comparisons! 

The user interface side is slightly more tricky. To specify a type to be 
booked, the user passes a Boost.Python.class like dataclasses.I3Particle. This 
gets packed in a struct called TypeSpec, which has a member function called 
check(I3FrameObjectPtr). When check(p) is called, p is converted to Python and 
compared to the held type via isinstance(). This will allow the user to specify 
that maps and vectors should be booked, which was simply not possible before. 
Hooray!

The unit/doc fields are also now written to the HDF file. They are stored as 
attributes e.g. FIELD_0_UNIT and FIELD_0_DOC, similarly to the field names. 
Hooks for pytables to access this information are in the works.  


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61500 16731396-06f5-0310-8873-f7f720988828