Added reading support to I3HDFTable.
I3Table now has a reading interface as well. This could be handy for
verification purposes, or for writing HDF5->other format converters after the
fact. Instead of reprocessing the data or implementing a format converter e.g.
in PyTables+PyROOT, you simply chain a reader and writer service together.
Some additional glue logic (including proper indexing) is required to make
this work, though.

The default implementation simply raises an exception, so read support is
optional in subclasses. Reading support is implemented in I3HDFTableService
and I3HDFTable. Tests are included to ensure that I3TableRow and
I3TableRowDescription can be resurrected from HDF5 files without any nasty
surprises*.

* There are, naturally, some nasty surprises. To whit: it's not always
possible to infer exactly which type a converter used to define a field. The
cardinal example is bool, which is represented as a char on some systems and
int on others. When the file is read back, the field type will be that of the
native representation, e.g. on Darwin 4 bytes rather than 1. When the buffers
are read out, however, they are identical, bit for bit.

The same issue makes it difficult to invert the mapping from a native type to
a python-style type code. Since the only thing stored on disk is the HDF5
representation, it's impossible to know whether the user was trying to write
out a bool or an int. For this reason, the python type codes are not
reconstructed when the I3TableRowDescription is read back from disk. I'm
working on a fix for this.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61565 16731396-06f5-0310-8873-f7f720988828