Added self-consistent handling of bool values.
Since C has no core boolean type distinct from the integers, we have to handle
this as a special case. I3TableRowDescription::AddField<bool>() now forces the
user to set "bool" as the unit string. This is justifiable, as the boolean
interpretation of integers is a convention, and conventions of interpretation
are defined by units. The Python-style typecode 'o' is now strictly tied to
'bool' being set as the unit, so it's become redundant.
I3TableRowDescription::Set<bool>() now explicitly casts the passed value to a
boolean via value!=false.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61572 16731396-06f5-0310-8873-f7f720988828