More indexing/multi-row object support.
Padding is now restricted to non-ragged tables, as it would make no sense to
pad out a table with variable numbers of rows per frame.
I3Converter::GetNumberOfRows() is now called before a table is instantiated,
so that the converter always has a "live" object with which to construct its
description. Marking a description with I3TableRowDescription::SetIsMultiRow()
is now mandatory; I3Table checks this every time a multi-row block is added to
the table.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61608 16731396-06f5-0310-8873-f7f720988828