Added a method to easily extend converters from userland.
Added I3ConverterBundle, a subclass of I3Converter that holds a collection of
converters. Its description is the concatenation of the descriptions provided
by its component converters, and its Convert() method calls the Convert()
methods of its component converters in turn. This is a generalization of the
way I3TableWriter inserts header information into every row written to a
table.

To create bundle, the user simply passes a list of converters instead of a
single I3Converter instance. This allows the user to book additional
information without having to subclass specific converters, and lets us keep
flexibility without placing too many requirements on converters define in
other projects.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/new-internals-2@61662 16731396-06f5-0310-8873-f7f720988828