Support table types which are always aligned by construction.
Added a virtual function I3Table::GetAlignmentType() which defaults to 
MultiRow. A derived class can instead return Strict. In this case, missing rows 
will always be padded out, regardless of whether or not the 
I3TableRowDescription sets the IsMultiRow flag. Previously, padding was not 
added for multi-row objects, as these could never be aligned via padding.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/hdf-writer/branches/tableio@61757 16731396-06f5-0310-8873-f7f720988828