Fixed a bug that arose when mixing single-row and multi-row converters with the same frame object, pointed out by @olle.
I3TableWriter refuses to instantiate a converter if it reports zero rows for a frame object (like I3RecoPulseSeriesMapConverter does for an empty pulse series), and defers until the next frame. This mechanism breaks down when another converter (like I3EventInfoConverterFromRecoPulses) reports at least one row for an object. Instantiation is then staggered, which caused delayed instantiations to be ignored.

Any TableSpecs still in wantedNames_ have not yet been instantiated, so we walk through this instead of checking for the frame key in tables_.

git-svn-id: http://code.icecube.wisc.edu/svn/projects/tableio/trunk@65296 16731396-06f5-0310-8873-f7f720988828