Fix a bug that broke row alignment when I3EventHeaders were changed by other modules.
Just because _we_ hold an I3EventHeaderConstPtr doesn't mean that someone
else isn't monkeying with it behind the scenes. In cases where there's no
event header (DST), it may be convenient to keep a potemkin I3EventHeader
around, incrementing its EventID and placing it in every header-less
frame that goes by. This makes for  very bad joo-joo, as there will only
be one entry in I3TableService::eventHeaderCache_, and that entry will in
point to the same instance as the I3EventHeaderPtr passed in which each
new frame, which is of course identical to itself. To prevent this sort
of thing from happening, we freeze a copy of the header before caching it.

This whole mechanism is kind of a hack; perhaps we should keep a running
index instead?


git-svn-id: http://code.icecube.wisc.edu/svn/projects/tableio/trunk@66289 16731396-06f5-0310-8873-f7f720988828